BAEL-1212 Guide to Creational Patterns - Made changes as per latest review
