--- core-java/src/main/java/com/baeldung/creationaldp/abstractfactory/AbstractFactory.java
@@ -1,6 +1,0 @@
-package com.baeldung.creationaldp.abstractfactory;
-
-public interface AbstractFactory {
-    Toy getToy(String toyType) ;
-    Color getColor(String colorType);
-}
