--- core-java/src/main/java/com/baeldung/creationaldp/abstractfactory/AbstractPatternDriver.java
@@ -1,18 +1,0 @@
-package com.baeldung.creationaldp.abstractfactory;
-
-public class AbstractPatternDriver {
-    public static void main(String[] args) {
-        AbstractFactory abstractFactory;
-        
-        //creating a brown toy dog
-        abstractFactory = FactoryProvider.getFactory("Toy");
-        Toy toy = abstractFactory.getToy("Dog");
-        
-        abstractFactory = FactoryProvider.getFactory("Color");
-        Color color = abstractFactory.getColor("Brown");
-        
-        String result = "A " + toy.getType() + " with " + color.getColor() + " color " + toy.makeSound();
-        
-        System.out.println(result);
-    }
-}
