--- core-java/src/main/java/com/baeldung/creationaldp/abstractfactory/Brown.java
@@ -1,10 +1,0 @@
-package com.baeldung.creationaldp.abstractfactory;
-
-public class Brown implements Color {
-
-    @Override
-    public String getColor() {
-        return "brown";
-    }
-
-}
