--- core-java/src/main/java/com/baeldung/creationaldp/abstractfactory/Color.java
@@ -1,5 +1,0 @@
-package com.baeldung.creationaldp.abstractfactory;
-
-public interface Color {
-    String getColor();
-}
