--- core-java/src/main/java/com/baeldung/creationaldp/abstractfactory/Dog.java
@@ -1,15 +1,0 @@
-package com.baeldung.creationaldp.abstractfactory;
-
-public class Dog implements Toy {
-
-    @Override
-    public String getType() {
-        return "Dog";
-    }
-
-    @Override
-    public String makeSound() {
-        return "Barks";
-    }
-
-}
