--- core-java/src/main/java/com/baeldung/creationaldp/abstractfactory/Duck.java
@@ -1,15 +1,0 @@
-package com.baeldung.creationaldp.abstractfactory;
-
-public class Duck implements Toy {
-
-    @Override
-    public String getType() {
-        return "Duck";
-    }
-
-    @Override
-    public String makeSound() {
-        return "Squeks";
-    }
-
-}
