--- core-java/src/main/java/com/baeldung/creationaldp/abstractfactory/FactoryProvider.java
@@ -1,15 +1,0 @@
-package com.baeldung.creationaldp.abstractfactory;
-
-public class FactoryProvider {
-    public static AbstractFactory getFactory(String choice){
-        
-        if("Toy".equalsIgnoreCase(choice)){
-            return new ToyFactory();
-        }
-        else if("Color".equalsIgnoreCase(choice)){
-            return new ColorFactory();
-        }
-        
-        return null;
-    }
-}
