--- core-java/src/main/java/com/baeldung/creationaldp/abstractfactory/Toy.java
@@ -1,6 +1,0 @@
-package com.baeldung.creationaldp.abstractfactory;
-
-public interface Toy {
-    String getType();
-    String makeSound();
-}
