--- core-java/src/main/java/com/baeldung/creationaldp/abstractfactory/White.java
@@ -1,10 +1,0 @@
-package com.baeldung.creationaldp.abstractfactory;
-
-public class White implements Color {
-
-    @Override
-    public String getColor() {
-        return "White";
-    }
-
-}
