--- core-java/src/main/java/com/baeldung/creationaldp/factory/Heptagon.java
@@ -1,10 +1,0 @@
-package com.baeldung.creationaldp.factory;
-
-public class Heptagon implements Polygon {
-
-    @Override
-    public String getType() {
-        return "Heptagon";
-    }
-
-}
