--- core-java/src/main/java/com/baeldung/creationaldp/factory/Octagon.java
@@ -1,10 +1,0 @@
-package com.baeldung.creationaldp.factory;
-
-public class Octagon implements Polygon {
-
-    @Override
-    public String getType() {
-        return "Octagon";
-    }
-
-}
