--- core-java/src/main/java/com/baeldung/creationaldp/factory/Pentagon.java
@@ -1,10 +1,0 @@
-package com.baeldung.creationaldp.factory;
-
-public class Pentagon implements Polygon {
-
-    @Override
-    public String getType() {
-        return "Pentagon";
-    }
-
-}
