--- core-java/src/main/java/com/baeldung/creationaldp/factory/Polygon.java
@@ -1,5 +1,0 @@
-package com.baeldung.creationaldp.factory;
-
-public interface Polygon {
-    String getType();
-}
