--- core-java/src/main/java/com/baeldung/creationaldp/factory/PolygonFactory.java
@@ -1,22 +1,0 @@
-package com.baeldung.creationaldp.factory;
-
-public class PolygonFactory {
-    public Polygon getPolygon(int numberOfSides) {
-        if(numberOfSides == 3) {
-            return new Triangle();
-        }
-        if(numberOfSides == 4) {
-            return new Square();
-        }
-        if(numberOfSides == 5) {
-            return new Pentagon();
-        }
-        if(numberOfSides == 4) {
-            return new Heptagon();
-        }
-        else if(numberOfSides == 8) {
-            return new Octagon();
-        }
-        return null;
-    }
-}
