--- core-java/src/main/java/com/baeldung/creationaldp/factory/Square.java
@@ -1,10 +1,0 @@
-package com.baeldung.creationaldp.factory;
-
-public class Square implements Polygon {
-
-    @Override
-    public String getType() {
-        return "Square";
-    }
-
-}
