--- core-java/src/main/java/com/baeldung/creationaldp/factory/Triangle.java
@@ -1,10 +1,0 @@
-package com.baeldung.creationaldp.factory;
-
-public class Triangle implements Polygon {
-
-    @Override
-    public String getType() {
-        return "Triangle";
-    }
-
-}
