--- core-java/src/main/java/com/baeldung/creationaldp/singleton/Singleton.java
@@ -1,13 +1,0 @@
-package com.baeldung.creationaldp.singleton;
-
-public class Singleton  {    
-    private Singleton() {}
-    
-    private static class SingletonHolder {    
-        public static final Singleton instance = new Singleton();
-    }    
-
-    public static Singleton getInstance() {    
-        return SingletonHolder.instance;    
-    }    
-}
