+++ core-java/src/main/java/com/baeldung/designpatterns/creational/abstractfactory/AbstractFactory.java
@@ -1,0 +1,6 @@
+package com.baeldung.designpatterns.creational.abstractfactory;
+
+public interface AbstractFactory {
+    Toy getToy(String toyType) ;
+    Color getColor(String colorType);
+}
