+++ core-java/src/main/java/com/baeldung/designpatterns/creational/abstractfactory/AbstractPatternDriver.java
@@ -1,0 +1,18 @@
+package com.baeldung.designpatterns.creational.abstractfactory;
+
+public class AbstractPatternDriver {
+    public static void main(String[] args) {
+        AbstractFactory abstractFactory;
+        
+        //creating a brown toy dog
+        abstractFactory = FactoryProvider.getFactory("Toy");
+        Toy toy = abstractFactory.getToy("Dog");
+        
+        abstractFactory = FactoryProvider.getFactory("Color");
+        Color color = abstractFactory.getColor("Brown");
+        
+        String result = "A " + toy.getType() + " with " + color.getColor() + " color " + toy.makeSound();
+        
+        System.out.println(result);
+    }
+}
