+++ core-java/src/main/java/com/baeldung/designpatterns/creational/abstractfactory/Brown.java
@@ -1,0 +1,10 @@
+package com.baeldung.designpatterns.creational.abstractfactory;
+
+public class Brown implements Color {
+
+    @Override
+    public String getColor() {
+        return "brown";
+    }
+
+}
