+++ core-java/src/main/java/com/baeldung/designpatterns/creational/abstractfactory/Color.java
@@ -1,0 +1,5 @@
+package com.baeldung.designpatterns.creational.abstractfactory;
+
+public interface Color {
+    String getColor();
+}
