+++ core-java/src/main/java/com/baeldung/designpatterns/creational/abstractfactory/Dog.java
@@ -1,0 +1,15 @@
+package com.baeldung.designpatterns.creational.abstractfactory;
+
+public class Dog implements Toy {
+
+    @Override
+    public String getType() {
+        return "Dog";
+    }
+
+    @Override
+    public String makeSound() {
+        return "Barks";
+    }
+
+}
