+++ core-java/src/main/java/com/baeldung/designpatterns/creational/abstractfactory/Duck.java
@@ -1,0 +1,15 @@
+package com.baeldung.designpatterns.creational.abstractfactory;
+
+public class Duck implements Toy {
+
+    @Override
+    public String getType() {
+        return "Duck";
+    }
+
+    @Override
+    public String makeSound() {
+        return "Squeks";
+    }
+
+}
