+++ core-java/src/main/java/com/baeldung/designpatterns/creational/abstractfactory/FactoryProvider.java
@@ -1,0 +1,15 @@
+package com.baeldung.designpatterns.creational.abstractfactory;
+
+public class FactoryProvider {
+    public static AbstractFactory getFactory(String choice){
+        
+        if("Toy".equalsIgnoreCase(choice)){
+            return new ToyFactory();
+        }
+        else if("Color".equalsIgnoreCase(choice)){
+            return new ColorFactory();
+        }
+        
+        return null;
+    }
+}
