+++ core-java/src/main/java/com/baeldung/designpatterns/creational/abstractfactory/Toy.java
@@ -1,0 +1,6 @@
+package com.baeldung.designpatterns.creational.abstractfactory;
+
+public interface Toy {
+    String getType();
+    String makeSound();
+}
