+++ core-java/src/main/java/com/baeldung/designpatterns/creational/abstractfactory/White.java
@@ -1,0 +1,10 @@
+package com.baeldung.designpatterns.creational.abstractfactory;
+
+public class White implements Color {
+
+    @Override
+    public String getColor() {
+        return "White";
+    }
+
+}
