+++ core-java/src/main/java/com/baeldung/designpatterns/creational/factory/Heptagon.java
@@ -1,0 +1,10 @@
+package com.baeldung.designpatterns.creational.factory;
+
+public class Heptagon implements Polygon {
+
+    @Override
+    public String getType() {
+        return "Heptagon";
+    }
+
+}
