+++ core-java/src/main/java/com/baeldung/designpatterns/creational/factory/Octagon.java
@@ -1,0 +1,10 @@
+package com.baeldung.designpatterns.creational.factory;
+
+public class Octagon implements Polygon {
+
+    @Override
+    public String getType() {
+        return "Octagon";
+    }
+
+}
