+++ core-java/src/main/java/com/baeldung/designpatterns/creational/factory/Pentagon.java
@@ -1,0 +1,10 @@
+package com.baeldung.designpatterns.creational.factory;
+
+public class Pentagon implements Polygon {
+
+    @Override
+    public String getType() {
+        return "Pentagon";
+    }
+
+}
