+++ core-java/src/main/java/com/baeldung/designpatterns/creational/factory/Polygon.java
@@ -1,0 +1,5 @@
+package com.baeldung.designpatterns.creational.factory;
+
+public interface Polygon {
+    String getType();
+}
