+++ core-java/src/main/java/com/baeldung/designpatterns/creational/factory/Square.java
@@ -1,0 +1,10 @@
+package com.baeldung.designpatterns.creational.factory;
+
+public class Square implements Polygon {
+
+    @Override
+    public String getType() {
+        return "Square";
+    }
+
+}
