+++ core-java/src/main/java/com/baeldung/designpatterns/creational/factory/Triangle.java
@@ -1,0 +1,10 @@
+package com.baeldung.designpatterns.creational.factory;
+
+public class Triangle implements Polygon {
+
+    @Override
+    public String getType() {
+        return "Triangle";
+    }
+
+}
