+++ core-java/src/main/java/com/baeldung/designpatterns/creational/singleton/Singleton.java
@@ -1,0 +1,13 @@
+package com.baeldung.designpatterns.creational.singleton;
+
+public class Singleton  {    
+    private Singleton() {}
+    
+    private static class SingletonHolder {    
+        public static final Singleton instance = new Singleton();
+    }    
+
+    public static Singleton getInstance() {    
+        return SingletonHolder.instance;    
+    }    
+}
