+++ core-java/src/main/java/com/baeldung/designpatterns/creational/singleton/SingletonDriver.java
@@ -1,0 +1,8 @@
+package com.baeldung.designpatterns.creational.singleton;
+
+public class SingletonDriver {
+    public static void main(String[] args) {
+        Singleton instance = Singleton.getInstance();
+        System.out.println(instance.toString());
+    }
+}
